+++ CHANGES.md
@@ -2,6 +2,24 @@
 
 ## Version 0.7.X
 
+### Version 0.7.2 (2026-03-27)
+
+#### Fixes
+* `RadialSolver`: Fixed bug in Takeuchi starting conditions where y6 was pulling the incorrect value. Kamata starting conditions were not affected.
+
+#### Performance
+* The changes introduced in this release led to around 25% improvement in TidalPy's RadialSolver Love Number Calculator speed.
+
+#### Dependencies
+* TidalPy now works on Python 3.14.
+* Bumped version of `ipympl` to `<=0.11.0`.
+* Bumped version of `CyRK` to `>=0.17.1, <0.18.1`.
+* Increased numpy's max pinnings to `<2.5`.
+
+#### Other
+* Changed `TidalPy.constants.d_PI_DBL` to `TidalPy.constants.d_PI` since the "DBL" is redundant with the "d_".
+  * Same for `d_INF`, `d_EPS_10`, `d_EPS_100`, `d_EPS`, `d_NAN`
+
 ### Version 0.7.1 (2026-02-05)
 
 #### Fixes